fix: Added Comments and fixed issues with build
